requirements.txt
# ============================================================================
# MT5 Trading Bot - Full Requirements
# ============================================================================
# For minimal installation, use: pip install -r requirements-core.txt
# ============================================================================

# Data Analysis & Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Backtesting & Trading
zipline-reloaded>=3.0.0
exchange-calendars>=4.2.0

# Database & Storage
sqlalchemy>=2.0.0

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Statistical Analysis & Machine Learning
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# GUI (Optional - only needed for account setup GUI)
tkinter-tooltip>=2.1.0
ttkbootstrap>=1.10.1

# ============================================================================
# OPTIONAL DEPENDENCIES (Install separately as needed)
# ============================================================================

# MT5 Integration (Windows only, or Linux/Mac with Wine)
# pip install MetaTrader5>=5.0.45

# Technical Analysis (Requires system dependencies)
# Linux: sudo apt-get install ta-lib
# Mac: brew install ta-lib
# Windows: Download from https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib
# Then: pip install TA-Lib

# Note: tkinter is usually included with Python
# If missing: sudo apt-get install python3-tk (Linux)
